requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
meraki==1.33.0
multidict==6.0.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
urllib3==2.0.2
XlsxWriter==3.1.2
yarl==1.9.2
